Master william: confirm merge --so far no conflicts

+++ Insantani2/app/src/main/java/com/williamhenry/insantani/SettingsFragment.java
@@ -17,11 +17,26 @@ import android.widget.Button;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
 
+import com.android.volley.Request;
+import com.android.volley.RequestQueue;
+import com.android.volley.Response;
+import com.android.volley.VolleyError;
+import com.android.volley.toolbox.StringRequest;
+
+import org.json.JSONObject;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Map;
+
 
 public class SettingsFragment extends Fragment {
     private SharedPreferences pref;
     private Editor editor;
     private RelativeLayout relativeLayout;
+    private String url;
+    private RequestQueue mQueue;
+    public static final String REQUEST_TAG = "SettingsFragement";
 
 
     public SettingsFragment() {
@@ -45,13 +60,14 @@ public class SettingsFragment extends Fragment {
         final boolean checkRefreshToken= pref.contains("refresh_token");
         final boolean tokenType= pref.contains("token_type");
         final boolean user_id=pref.contains("user_id");
+        final boolean checkRegistrationToken=pref.contains("notification_token");
         Button logout= (Button) rootView.findViewById(R.id.logoutButton);
         relativeLayout=(RelativeLayout) rootView.findViewById(R.id.LogoutLayout);
         logout.setOnClickListener(new View.OnClickListener(){
             @Override
             public void onClick(View view) {
                 Log.d("token", Boolean.toString(checkToken));
-                if(!checkToken && !tokenType && !checkRefreshToken && !user_id ){
+                if(!checkToken && !tokenType && !checkRefreshToken && !user_id && !checkRegistrationToken){
 //                    logout.setEnabled(false);
 
                     Snackbar snackbar = Snackbar.make(relativeLayout, "Login First", Snackbar.LENGTH_LONG);
@@ -64,10 +80,61 @@ public class SettingsFragment extends Fragment {
                     snackbar.show();
                 }else{
                     editor=pref.edit();
-                    editor.clear();
-                    editor.commit();
-                    Intent intent= new Intent (getActivity(),MainActivity.class);
-                    startActivity(intent);
+
+                    mQueue= CustomVolleyRequestQueue.getInstance(getContext()).getRequestQueue();
+                    Log.d("notification_registration",pref.getString("notification_token",null));
+                    url="http://104.155.213.80/insantani/public/api/notify/register?user_id="+pref.getString("user_id",null)
+                            +"&device_token="+pref.getString("notification_token",null);
+                    final StringRequest stringRequestDeleteNotification= new StringRequest(Request.Method.DELETE,
+                            url, new Response.Listener<String>(){
+                        //                    private ArrayList<Article> articles1=new ArrayList<Article>();
+                        @Override
+                        public void onResponse(String response){
+                            Log.d("delete_notification_registration", response.toString());
+                            try {
+                                Log.d("response_delete_notification_registration", response.toString());
+                                JSONObject jsonObject= new JSONObject(response.toString());
+                                String message= jsonObject.getString("message");
+                                if(message.equals("Delete Success")) {
+                                    editor.clear();
+                                    editor.commit();
+                                    Intent intent= new Intent (getActivity(),MainActivity.class);
+                                    startActivity(intent);
+
+                                    Log.d("delete_notification_register_message",message);
+                                }
+
+                            } catch(Exception e){
+                                Log.d("JSON_error_delete_notification_register",e.toString());
+                            }
+                        }
+                    },new Response.ErrorListener(){
+                        @Override
+                        public void onErrorResponse(VolleyError error){
+
+                            Log.d("error_response_delete_notification_register",error.toString());
+
+                        }
+                    }){
+                        @Override
+                        public Map<String,String> getHeaders(){
+                            Map<String,String> headers= new HashMap<String, String>();
+                            String auth="Bearer "+pref.getString("access_token",null);
+                            Log.d("auth_notificaation_register", auth);
+                            headers.put("Authorization",auth);
+                            return headers;
+                        };
+                    };
+
+
+                    stringRequestDeleteNotification.setTag(REQUEST_TAG);
+                    mQueue.add(stringRequestDeleteNotification);
+
+
+                    //real logout
+
+
+
 
 //                    Snackbar snackbar = Snackbar.make(relativeLayout, "Logout", Snackbar.LENGTH_LONG);
 //                    snackbar.setActionTextColor(Color.WHITE);
